+++ pytorch_lightning/pt_overrides/__init__.py
@@ -3,7 +3,7 @@
  The deprecated module name will be removed in v0.8.0.
 """
 
-import warnings
+from pytorch_lightning.utilities import rank_zero_warn
 
-warnings.warn("`pt_overrides` package has been renamed to `overrides` since v0.6.0."
-              " The deprecated module name will be removed in v0.8.0.", DeprecationWarning)
+rank_zero_warn("`pt_overrides` package has been renamed to `overrides` since v0.6.0."
+               " The deprecated module name will be removed in v0.8.0.", DeprecationWarning)